+++ R/odbc-connection.R
@@ -20,7 +20,9 @@ OdbcConnection <- function(
   stopifnot(all(has_names(attributes)))
   stopifnot(all(names(attributes) %in% SUPPORTED_CONNECTION_ATTRIBUTES))
 
-  connection_string <- build_connection_string(..., .string = .connection_string)
+  args <- compact(list(...))
+  check_args(args)
+  connection_string <- build_connection_string(args, .connection_string)
 
   bigint <- bigint_mappings()[match.arg(bigint, names(bigint_mappings()))]
 
@@ -28,14 +30,19 @@ OdbcConnection <- function(
     timeout <- 0
   }
 
-  ptr <- odbc_connect(
-    connection_string,
-    timezone = timezone,
-    timezone_out = timezone_out,
-    encoding = encoding,
-    bigint = bigint,
-    timeout = timeout,
-    r_attributes_ = attributes
+  withCallingHandlers(
+    ptr <- odbc_connect(
+      connection_string,
+      timezone = timezone,
+      timezone_out = timezone_out,
+      encoding = encoding,
+      bigint = bigint,
+      timeout = timeout,
+      r_attributes_ = attributes
+    ),
+    error = function(cnd) {
+      check_quoting(args)
+    }
   )
   quote <- connection_quote(ptr)
 
@@ -63,22 +70,61 @@ OdbcConnection <- function(
   )
 }
 
-build_connection_string <- function(.string = NULL, ...) {
-  args <- compact(list(...))
-  check_args(args)
+build_connection_string <- function(args = list(), string = NULL) {
 
-  needs_escape <- grepl("[{}(),;?*=!@]", args) &
-    !grepl("^\\{.*\\}$", args) &
-    !vapply(args, inherits, "AsIs", FUN.VALUE = logical(1))
-
-  args[needs_escape] <- paste0("{", args[needs_escape], "}")
   args_string <- paste(names(args), args, sep = "=", collapse = ";")
 
-  if (!is.null(.string) && !grepl(";$", .string) && length(args) > 0) {
-    .string <- paste0(.string, ";")
+  if (!is.null(string) && !grepl(";$", string) && length(args) > 0) {
+    string <- paste0(string, ";")
   }
 
-  paste0(.string, args_string)
+  paste0(string, args_string)
+}
+
+#' Quote special character when connecting
+#'
+#' @description
+#' When connecting to a database using odbc, all the arguments are concatenated
+#' into a single connection string that looks like `name1=value1;name2=value2`.
+#' That means if your value contains `=` or `;` then it needs to be quoted.
+#' Other rules mean that you need to quote any values that starts or ends with
+#' white space, or contains `{` or `}`.
+#'
+#' This function quotes a string in a way that should work for most drivers,
+#' but unfortunately there doesn't seem to be an approach that works everywhere.
+#' If this function doesn't work for you, you'll need to carefully read the
+#' docs for your driver.
+#'
+#' @export
+#' @param x A string to quote.
+#' @return A quoted string, wrapped in `I()`.
+#' @examples
+#' quote_value("abc")
+#' quote_value("ab'c")
+#'
+#' # Real usage is more likely to look like:
+#' \dontrun{
+#' library(DBI)
+#'
+#' con <- dbConnect(
+#'   odbc::odbc(),
+#'   dsn = "reallycooldatabase"
+#'   password = odbc::quote_value(Sys.getenv("MY_PASSWORD"))
+#' )
+#' }
+quote_value <- function(x) {
+  has_single <- grepl("'", x, fixed = TRUE)
+  has_double <- grepl('"', x, fixed = TRUE)
+
+  if (has_single && has_double) {
+    abort("Don't know how to escape a value with both single and double quotes.")
+  } else if (has_double) {
+    quote <- "'"
+  } else {
+    quote <- '"'
+  }
+
+  I(paste0(quote, x, quote))
 }
 
 check_args <- function(args) {
@@ -103,6 +149,33 @@ check_args <- function(args) {
       call = quote(DBI::dbConnect())
     )
   }
+}
+
+check_quoting <- function(args) {
+  needs_quoting <- vapply(args, needs_quoting, FUN.VALUE = logical(1))
+  if (any(needs_quoting)) {
+    # TODO: use cli pluralisation
+    args <- paste0("`", names(args)[needs_quoting], "`", collapse = ", ")
+
+    inform(c(
+      paste0(args, " contains a special character that may need quoting."),
+      i = "Wrap the value in `odbc::quote_value()` to use a heuristic that should work for most drivers.",
+      i = "If that still doesn't work, consult your driver's documentation.",
+      i = "Otherwise, you can suppress this message by wrapping the value in `I()`."
+    ))
+  }
+}
+
+needs_quoting <- function(x) {
+  if (inherits(x, "AsIs")) {
+    return(FALSE)
+  }
+
+  if (grepl("^\\{.*\\}$", x) || grepl("^(['\\\"]).*\\1$", x)) {
+    return(FALSE)
+  }
+
+  grepl("[{}=;]| $|^ ", x)
 }
 
 # -------------------------------------------------------------------------
@@ -140,16 +213,38 @@ NULL
 
 #' odbcConnectionColumns
 #'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' This function has been deprecated in favor of [dbListFields()].
+#'
 #' For a given table this function returns detailed information on
 #' all fields / columns.  The expectation is that this is a relatively thin
 #' wrapper around the ODBC `SQLColumns` function call, with some of the field
 #' names renamed / re-ordered according to the return specifications below.
 #'
+#' @details
 #' In [dbWriteTable()] we make a call to this method
 #' to get details on the fields of the table we are writing to.  In particular
 #' the columns `data_type`, `column_size`, and `decimal_digits` are used.  An
 #' implementation is not necessary for [dbWriteTable()] to work.
-#' @inheritParams odbcConnectionTables
+#'
+#' `odbcConnectionColumns` is routed through the `SQLColumns` ODBC
+#' method.  This function, together with remaining catalog functions
+#' (`SQLTables`, etc), by default ( `SQL_ATTR_METADATA_ID == false` ) expect
+#' either ordinary arguments (OA) in the case of the catalog, or pattern value
+#' arguments (PV) in the case of schema/table name.  For these, quoted
+#' identifiers do not make sense, so we unquote identifiers prior to the call.
+#'
+#' @param conn OdbcConnection
+#' @param name Table identifiers.
+#'   By default, are interpreted as a ODBC search pattern where `_` and `%` are
+#'   wild cards. Set `exact = TRUE` to match `_` exactly.
+#' @param ... additional parameters to methods
+#' @param exact Set to `TRUE` to escape `_` in identifier names so that it
+#'   matches exactly, rather than matching any single character. `%` always
+#'   matches any number of characters as this is unlikely to appear in a
+#'   table name.
 #'
 #' @seealso The ODBC documentation on
 #' [SQLColumns](https://docs.microsoft.com/en-us/sql/odbc/reference/syntax/sqlcolumns-function)
@@ -175,22 +270,29 @@ NULL
 #' - nullable
 #'
 #' @rdname odbcConnectionColumns
+#' @keywords internal
 #' @export
+odbcConnectionColumns <- function(conn, name, ..., exact = FALSE) {
+  lifecycle::deprecate_warn(
+    "1.4.2",
+    "odbcConnectionColumns()",
+    "dbListFields()"
+  )
+
+  odbcConnectionColumns_(conn = conn, name = name, ..., exact = exact)
+}
+
 setGeneric(
-  "odbcConnectionColumns",
+  "odbcConnectionColumns_",
   valueClass = "data.frame",
   function(conn, name, ..., exact = FALSE) {
-    standardGeneric("odbcConnectionColumns")
+    standardGeneric("odbcConnectionColumns_")
   }
 )
 
-#' @rdname odbcConnectionColumns
-#' @param column_name The name of the column to return, the default returns
-#'   all columns.
-#' @export
-setMethod("odbcConnectionColumns", c("OdbcConnection", "Id"),
+setMethod("odbcConnectionColumns_", c("OdbcConnection", "Id"),
   function(conn, name, ..., column_name = NULL, exact = FALSE) {
-    odbcConnectionColumns(conn,
+    odbcConnectionColumns_(conn,
       name = id_field(name, "table"),
       catalog_name = id_field(name, "catalog"),
       schema_name = id_field(name, "schema"),
@@ -200,9 +302,7 @@ setMethod("odbcConnectionColumns", c("OdbcConnection", "Id"),
   }
 )
 
-#' @rdname odbcConnectionColumns
-#' @export
-setMethod("odbcConnectionColumns", c("OdbcConnection", "character"),
+setMethod("odbcConnectionColumns_", c("OdbcConnection", "character"),
   function(conn,
            name,
            ...,
@@ -225,59 +325,26 @@ setMethod("odbcConnectionColumns", c("OdbcConnection", "character"),
   }
 )
 
-#' @details `odbcConnectionColumns` is routed through the `SQLColumns` ODBC
-#'  method.  This function, together with remaining catalog functions
-#'  (`SQLTables`, etc), by default ( `SQL_ATTR_METADATA_ID == false` ) expect
-#'  either ordinary arguments (OA) in the case of the catalog, or pattern value
-#'  arguments (PV) in the case of schema/table name.  For these, quoted
-#'  identifiers do not make sense, so we unquote identifiers prior to the call.
-#' @seealso The ODBC documentation on
-#' [Arguments to catalog functions](https://learn.microsoft.com/en-us/sql/odbc/reference/develop-app/arguments-in-catalog-functions).
-#' @rdname odbcConnectionColumns
-#' @export
-setMethod("odbcConnectionColumns", c("OdbcConnection", "SQL"),
+setMethod("odbcConnectionColumns_", c("OdbcConnection", "SQL"),
   function(conn, name, ...) {
-    odbcConnectionColumns(conn, dbUnquoteIdentifier(conn, name)[[1]], ...)
+    odbcConnectionColumns_(conn, dbUnquoteIdentifier(conn, name)[[1]], ...)
   }
 )
 
-#' odbcConnectionTables
-#'
-#' This function returns a listing of tables accessible
-#' to the connected user.
-#' The expectation is that this is a relatively thin
-#' wrapper around the ODBC `SQLTables` function call,
-#' albeit returning a subset of the fields.
-#'
-#' It is important to note that, similar to the ODBC/API
-#' call, this method also accomodates pattern-value arguments
-#' for the catalog, schema, and table name arguments.
-#'
-#' If extending this method, be aware that `package:odbc`
-#' internally uses this method to satisfy both
-#' DBI::dbListTables and DBI::dbExistsTable methods.
-#' ( The former also advertises pattern value arguments )
-#'
-#' @param conn OdbcConnection
-#' @param name,catalog_name,schema_name Catalog, schema, and table identifiers.
-#'   By default, are interpreted as a ODBC search pattern where `_` and `%` are
-#'   wild cards. Set `exact = TRUE` to match `_` exactly.
-#' @param ... additional parameters to methods
-#' @param exact Set to `TRUE` to escape `_` in identifier names so that it
-#'   matches exactly, rather than matching any single character. `%` always
-#'   matches any number of characters as this is unlikely to appear in a
-#'   table name.
-#'
-#' @seealso The ODBC documentation on
-#' [SQLTables](https://docs.microsoft.com/en-us/sql/odbc/reference/syntax/sqlcolumns-function)
-#' for further details.
-#'
-#' @return data.frame with columns
-#' - table_catalog
-#' - table_schema
-#' - table_name
-#' - table_remarks
-#' @rdname odbcConnectionTables
+# This function returns a listing of tables accessible
+# to the connected user.
+# The expectation is that this is a relatively thin
+# wrapper around the ODBC `SQLTables` function call,
+# albeit returning a subset of the fields.
+#
+# It is important to note that, similar to the ODBC/API
+# call, this method also accomodates pattern-value arguments
+# for the catalog, schema, and table name arguments.
+#
+# If extending this method, be aware that `package:odbc`
+# internally uses this method to satisfy both
+# DBI::dbListTables and DBI::dbExistsTable methods.
+# ( The former also advertises pattern value arguments )
 setGeneric(
   "odbcConnectionTables",
   valueClass = "data.frame",
@@ -286,10 +353,6 @@ setGeneric(
   }
 )
 
-#' @rdname odbcConnectionTables
-#' @param table_type List tables of this type, for example 'VIEW'.
-#' See odbcConnectionTableTypes for a listing of available table
-#' types for your connection.
 setMethod("odbcConnectionTables", c("OdbcConnection", "Id"),
   function(conn, name, table_type = NULL, exact = FALSE) {
     odbcConnectionTables(
@@ -303,7 +366,6 @@ setMethod("odbcConnectionTables", c("OdbcConnection", "Id"),
   }
 )
 
-#' @rdname odbcConnectionTables
 setMethod("odbcConnectionTables", c("OdbcConnection", "character"),
   function(conn,
            name,
@@ -326,7 +388,6 @@ setMethod("odbcConnectionTables", c("OdbcConnection", "character"),
   }
 )
 
-#' @rdname odbcConnectionTables
 setMethod("odbcConnectionTables", "OdbcConnection",
   function(conn,
            name = NULL,
@@ -345,7 +406,6 @@ setMethod("odbcConnectionTables", "OdbcConnection",
   }
 )
 
-#' @rdname odbcConnectionTables
 setMethod("odbcConnectionTables", c("OdbcConnection", "SQL"),
   function(conn, name, table_type = NULL, exact = FALSE) {
     odbcConnectionTables(
@@ -357,12 +417,6 @@ setMethod("odbcConnectionTables", c("OdbcConnection", "SQL"),
   }
 )
 
-#' odbcConnectionCatalogs
-#'
-#' This function returns a listing of available
-#' catalogs.
-#' @param conn OdbcConnection
-#' @rdname odbcConnectionCatalogs
 setGeneric(
   "odbcConnectionCatalogs",
   valueClass = "character",
@@ -371,26 +425,14 @@ setGeneric(
   }
 )
 
-#' @rdname odbcConnectionCatalogs
 setMethod("odbcConnectionCatalogs", "OdbcConnection",
   function(conn) {
     connection_sql_catalogs(conn@ptr)
   }
 )
 
-#' odbcConnectionSchemas
-#'
-#' This function returns a listing of available
-#' schemas.
-#'
-#' Currently, for a generic connection the
-#' catalog_name argument is ignored.
-#'
-#' @param conn OdbcConnection
-#' @param catalog_name Catalog where
-#' we are looking to list schemas.
-#' @keywords internal
-#' @rdname odbcConnectionSchemas
+# Currently, for a generic connection the
+# catalog_name argument is ignored.
 setGeneric(
   "odbcConnectionSchemas",
   valueClass = "character",
@@ -399,19 +441,12 @@ setGeneric(
   }
 )
 
-#' @rdname odbcConnectionSchemas
 setMethod("odbcConnectionSchemas", "OdbcConnection",
   function(conn, catalog_name = NULL) {
     connection_sql_schemas(conn@ptr)
   }
 )
 
-#' odbcConnectionTableTypes
-#'
-#' This function returns a listing of table
-#' types available in database.
-#' @param conn OdbcConnection
-#' @rdname odbcConnectionTableTypes
 setGeneric(
   "odbcConnectionTableTypes",
   valueClass = "character",
@@ -420,7 +455,6 @@ setGeneric(
   }
 )
 
-#' @rdname odbcConnectionTableTypes
 setMethod("odbcConnectionTableTypes", "OdbcConnection",
   function(conn) {
     connection_sql_table_types(conn@ptr)